+++ packages/go_router/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 10.2.0
+
+- Adds `onExit` to GoRoute.
+
+## 10.1.4
+
+- Fixes RouteInformationParser that does not restore full RouteMatchList if
+  the optionURLReflectsImperativeAPIs is set.
+
 ## 10.1.3
 
 - Fixes an issue in the documentation that was using `state.queryParameters` instead of `state.uri.queryParameters`.
@@ -72,7 +81,7 @@
 
 - Makes namedLocation and route name related APIs case sensitive.
 
-## 8.0.2 
+## 8.0.2
 
 - Fixes a bug in `debugLogDiagnostics` to support StatefulShellRoute.